Adding new capabilities to IntCodeComputer; relative addressing and more memory

+++ src/year2019/day05.rs
@@ -2,13 +2,13 @@ use crate::error::AdventError;
 
 use super::int_code_computer::IntCodeComputer;
 
-type ParsedInput = Vec<i32>;
+type ParsedInput = Vec<i64>;
 
 pub fn parse(input: &str) -> color_eyre::Result<ParsedInput> {
     IntCodeComputer::parse_program(input)
 }
 
-pub fn part1(input: &ParsedInput) -> color_eyre::Result<i32> {
+pub fn part1(input: &ParsedInput) -> color_eyre::Result<i64> {
     let mut icc = IntCodeComputer::load(input.clone());
     icc.push_input(1);
     icc.run()?;
@@ -21,7 +21,7 @@ pub fn part1(input: &ParsedInput) -> color_eyre::Result<i32> {
         .cloned()
 }
 
-pub fn part2(input: &ParsedInput) -> color_eyre::Result<i32> {
+pub fn part2(input: &ParsedInput) -> color_eyre::Result<i64> {
     let mut icc = IntCodeComputer::load(input.clone());
     icc.push_input(5);
     icc.run()?;
